compiler file separate from parser
